folder structure refactored
